Week2/Multimodel-Rag_project/requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.0

# LLM Providers
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
groq==0.4.2

# Vector Database & Embeddings
faiss-cpu==1.8.0
sentence-transformers==2.3.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10

# Image Processing
opencv-python==4.9.0.80

# RAG & Search
rank-bm25==0.2.2

# Knowledge Graph
networkx==3.2.1
spacy==3.7.2

# Web Search
requests==2.31.0
beautifulsoup4==4.12.3

# Guardrails
detoxify==0.5.2

# Observability (Langsmith)
langsmith==0.0.87

# Database for Memory
sqlalchemy==2.0.25

# Utilities
pydantic==2.6.1
tiktoken==0.8.0
numpy==1.26.3
pandas==2.2.0
tenacity==8.2.3

# Knowledge Graph
spacy==3.7.2
networkx==3.2.1

# Hybrid Search
rank-bm25==0.2.2